Modify storage unit to generate a  new IV
Currently, the storage unit reuses the same IV each time a record
is stored.  This works (probably) for DES3, but not for AES.

The getWrappingParams() method is modified to check the config for
the special string "generate:X" where X is the number of bytes to be
generated in the IV.

To ensure that the same IV used to encrypt the secret is stored in the
DB, the wrapping param is defined once in the archival process, and
passed in to the wrapping functions in storageUnit.

Change-Id: Ia6696adf56fc7a4e90f83948c7549b64a38ab854