Added python info client
Add python client code to read from the InfoResource class and get
the server version.  As the PKIConnection in the python client
currently requires a subsystem, it is difficult to add an infoclient
to an existing KRAClient (or any other client).

To get around this, I modified the PKIConnection to allow using the
rootURI.

Change-Id: Ided75f45f741e2ba3fc86acec715d24b829c8a97