Update patch set 4
Patch Set 4: Code-Review-1

My only objection here is the checking in of commented out (and hence dead code.  In this case, that results in extra code and classes being checked on on both the server and client side.

If we really think we're going to implement this sometime, open a ticket (for 10.5) and attach the patches there.  Or even refer to this review - which has the patches already attached.

Then create a clean patch with the extraneous classes removed.

Patch-set: 4
Label: Code-Review=-1