KRA: use AES in PKCS #12 recovery for encrypted keys
The KRA has two private key recovery code paths: one dealing with
keys wrapped to the storage key, and one dealing with symmetrically
encrypted keys.  Each has a separate function for constructing a
PKCS #12 file for the recovered key.

This commit updates the PKCS #12 generation for encrypted keys to
use AES encryption.  From the KRA recovery process we start with a
byte[] of PrivateKeyInfo.  The previous procedure involved
encrypting that data, and manually building up the
EncryptedPrivateKeyInfo structure using JSS ASN.1 classes.  The new
procedure does away with all of this, and instead imports the
private key as a *temporary* key on the *internal* NSS software
token, then calls the same createPFX method used for wrapped keys.

The assumption is that the internal token is always available, and
appropriate to use for this purpose.  It is also assumed to work in
FIPS mode.  These assumptions need to be carefully checked.

Fixes: https://pagure.io/dogtagpki/issue/2610
Change-Id: Ie292147caab357679b2be5cf3b6cd739e5bed8e0