KRA: use AES in PKCS #12 recovery for encrypted keys
The KRA has two private key recovery code paths: one dealing with
keys wrapped to the storage key, and one dealing with symmetrically
encrypted keys.  Each has a separate function for constructing a
PKCS #12 file for the recovered key.

This commit updates the PKCS #12 generation for encrypted keys to
use AES encryption.  From the KRA recovery process we start with a
byte[] of PrivateKeyInfo.  The previous procedure used
EncryptedPrivateKey.createPBE() with PBEAlgorithm.PBE_SHA1_DES3_CBC
to encrypt the data.  This commit changes the procedure to use AES.

For PKCS #12 PBE algorithms the cipher was implied, but to use AES
algorithms, we must use PBES2, under which the KDF and cipher are
configured separately; the PBE algorithm does not imply the cipher
(therefore does not imply the key length).  So we can't use
EncryptedPrivateKeyInfo.createPBE() because there is no way to
convey the intended encryption algorithm or the required key length
to the symmetric key generation machinery.

Instead, perform the key derivation directly via KeyGenerator, then
encipher the private key data, and finally construct the PBES2
parameters and EncryptedPrivateKeyInfo.

Fixes: https://pagure.io/dogtagpki/issue/2664

Change-Id: Ie292147caab357679b2be5cf3b6cd739e5bed8e0