Add field to indicate if key was encrypted or wrapped
Whether a secret was encrypted or wrapped in the storage unit
depends on a parameter in CS.cfg.  If that parameter is changed,
the Storage unit may use the wrong mechanism to try to decrypt
the stored key.  Thats ok for encrypt/wrap using DES or AES-CBC,
but not for AES KeyWrap.

In this patch, we add a field in the Key record to specify whether
the secret was encrypted with stored (or keywrapped if false).

A subsequent patch will change the logic when decrypting to use
this field.

Change-Id: If535156179bd1259cfaaf5e56fd4d36ffdb0eb0e