Allow key recovery to use encrypted field in key record
The previous commit added a field in the KeyRecord to
specify whether or not a key was encrypted or key wrapped
when archived.  This patch modifies the recovery servlets
to use this field to determine how to decrypt/unwrap the
key for transport.

Absence of this field in the key record implies that is
an old record - and we use the value of the CS.cfg parameter
as the default.

Change-Id: Ia8ae679e8b3fe8462d42848d614bff863ef68e50