Refactored CertRequestProcessedEvent to use AuditEvent attributes.
The CertRequestProcessedEvent constructors have been modified to
log the info attributes using the new AuditEvent attributes.

The logging property for CERT_REQUEST_PROCESSED event has been
modified to accept a list of attributes as a single string instead
of individual info attributes.

The CERT_REQUEST_PROCESSED constant in AuditEvent has been replaced
with a constant in CertRequestProcessedEvent class which points to
the new logging property.

https://pagure.io/dogtagpki/issue/2655

Change-Id: I981212af7fca58916c73ccdeba9919a4d051af3c