Split CERT_REQUEST_PROCESSED event.
The LogMessages.properties has been modified to remove the old
CERT_REQUEST_PROCESSED event and replace it with the new
CERT_REQUEST_FAILURE and CERT_REQUEST_SUCCESS events.

The CertRequestFailureEvent and CertRequestSuccessEvent have been
modified to use the new logging properties.

The CS.cfg in all subsystems have also been modified to replace
the old event with the new ones.

https://pagure.io/dogtagpki/issue/2647

Change-Id: I181de7b8008c0eb825b384288cc10036f4893f80