Update patch set 1
Patch Set 1:

I'm ambivalent about the need for this.  Remember that the default that is returned in case the KRA cannot be contacted is 3DES which is supposed to work on either old or new servers.

Not sure if the client cares if the response is authoritative or not.

Patch-set: 1
Label: Workflow=0