Update patch set 1
Patch Set 1: Code-Review+2

Looks good. One quick thing, if you have not done so, is to generate and recover a key and make sure the key is the same somehow. Maybe encrypt some data with it such as a zero vector and see if the result is the same for both. ACK though.

Patch-set: 1
Reviewer: Jack Magne <1010222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2