Fixed CERT_REQUEST_PROCESSED events in ConnectorServlet.
The code that generates CERT_REQUEST_PROCESSED events in
ConnectorServlet.processRequest() has been moved into a finally-
clause that wraps around IRequestQueue.processRequest() to ensure
that the events are generated properly.

If a cert was issued for the request that has just been processed
the event outcome is a Success, otherwise it's a Failure.

Any exception thrown by the IRequestQueue.processRequest() will be
passed to the ConnectorServlet.processRequest()'s callers.

https://pagure.io/dogtagpki/issue/2690

Change-Id: I07454afb75328fbee3e50e5852adb5085be0613e