Fix failing audit log
As currently written, the audit log for completing the cert
processing on the KRA will always fail because the cert is not
yet issued.  The cert is only issued after the key is archived.

In this patch, we simply replace the failed audit log with a success
one to indicate that the operation has been completed.

Change-Id: I9e658dca15fd87e87c0124c4c9972dbca2910643