Fix failing audit log
As currently written, the audit log for completing the cert
processing on the KRA will always fail because the cert is not
yet issued.  The cert is only issued after the key is archived.

Basically, though, this particular log is only suppposed to be
written to the CA audit log.  Rather than adding a subsystem check,
the simplest solution is to not expose this event on the KRA.

Change-Id: I9e658dca15fd87e87c0124c4c9972dbca2910643