Make sure archivalID is passed through archival
There was some confusion in the previous commit for archival
logging.  The archivalID is the id provided by the CA for the archival
and is its requestID.  This allows the cert request operation
to be tracked through the archival.

Made sure therefore, that we have two fields - one for the archivalID
and one for the requestId (which is the KRA archival request ID)

In addition, some of the archival events occur in the CA component
just before the request id sent to the KRA.  These events will not
be displayed unless the audit event is added to the CA CS.cfg.

Change-Id: I3904d42ae677d5916385e0120f0e25311b4d9d08