Simplify recovery audit logging
Currently, when we use the retrieveKey() REST interface, there are
two logs generated for the processing of a recovery request.  To
rectify this, logging has been removed from the lower level in the
SecurityDataProcessor and is delegated to the higher level.

This necessitated adding audit logging to the SecurityDataRecoveryService,
which processes recovery events asynchronously.

In addition, the logging in retrieveKey() has been pushed down to
the retrieveKeyImpl, because there is at least one success exit point in
retrieveKeyImpl where a recovery request is created, but no key is exported.
Hence in this case, a KeyRetrieve success event is not warranted.

Change-Id: I0725e6fe82046ae666bf6c81d6a6ba58261dfc87