Update patch set 1
Patch Set 1:

Actually the same OCSP responder processes all OCSP requests;
the CA subsystem's internal OCSP responder will look up which
authority issued the cert and use the same authority to sign
the OCSP response.

So I guess the main need for this command and the work it does
retrieving the CA certificate is so that the OCSP response can be validated.  Is that right?

Patch-set: 1
Label: Workflow=0