Refactored temp SSL server cert creation (part 2).
The create_temp_sslserver_cert() in configuration.py has been
modified to check for SSL server cert in the NSS database first.
If there is an existing SSL server cert and the hostname in the
subject DN matches the hostname of the server, it will reuse the
existing cert. Otherwise, it will remove the existing cert and its
key, and generate a temp cert as before.

The pki_self_signed_nickname has been removed since it's identical
to pki_ssl_server_nickname.

https://pagure.io/dogtagpki/issue/2786

Change-Id: I461adef85301afa199dab49f6abad4ca709ffd8c