Fixed pki-server cert-find output.
The pki-server cert-find output has been modified such that it
is more consistent with the pki ca-cert-find output.

The serial number is now stored as an integer and displayed in
hexadecimal format.

If the certificate does not exist yet in the NSS database, the
attributes that depend on it (e.g. serial number) will not be
displayed. This allows the tool to be used during installation
to inspect the content of the NSS database.

https://pagure.io/dogtagpki/issue/2776

Change-Id: Id99806ab750a51dc31e9e8f32c35b8f5e2bf64e1