Refactored key generation for SSL server certificate.
The SystemConfigService has been modified not to generate a new
key if there is already an existing key in the NSS database,
either generated earlier in configuration.py or imported from
another instance.

For SSL server certificate this means that the key created for
the temporary certificate will be reused for the permanent
certificate.

https://pagure.io/dogtagpki/issue/2280

Change-Id: Ia972e1df08a3c76b779f497dd3b47a0a50330a0b