Added generic CMC servlet.
The caProfileSubmitCMCFull servlet has been modified to accept
the profile ID specified in the URL query parameter, which will
override the default profile ID specified in the web.xml.

The HttpClient was previously using StringTokenizer with '=' as
delimiter to parse the configuration file, which caused a problem
when parsing URL query parameters. The code has been modified to
use the Properties class to parse the configuration file properly.

https://pagure.io/dogtagpki/issue/2815

Change-Id: I99c63bcef7c9e6975c5e8652693c6467c5424ae6