Fix pki-server subsystem-cert-validate command with big serial
The pki-server subsystem-cert-validate command fails when dealing
with a certificate with a large serial number. This can cause
installation failure.

The problem arises because pki.nssdb.NSSDatabase.get_cert_info()
uses a regular expression to search for the serial number in the
certutil -L (pretty-print) output.  When the serial number is small
it is formatted like:

        ...
        Serial Number: 15 (0xf)
        ...

But when the serial number is large, it is formatted like:

        ...
        Serial Number:
            48:00:00:00:11:04:b6:f1:74:bd:63:2c:4b:00:00:00
            00:00:11
        ...

This causes the regex to fail to fail, crashing the program.

Use python-cryptography to parse the cert and read the serial number
instead of interpreting the pretty-printed cert.

This begs a subsequent refactor to read the other attributes from
the python-cryptography Certificate object but this is left for
another day.

Fixes: https://pagure.io/dogtagpki/issue/2827