Update patch set 1
Patch Set 1:

Any idea why the preferred nickname was never used?

I think the current code was assuming the first cert will be the leaf cert and it will be imported using the preferred nickname. Your patch assumes that the leaf cert will be the last.

IIRC, according to cfu the order of the certificates in the PKCS #7 is not guaranteed. So we probably will need to do this:

* sort the certificates based on the subject DNs & issuer DNs (we probably can't use buildCertificateChain() since we'd have to provide the leaf cert, which we don't have yet at this point)
* trust the leaf cert with the trust flags provided
* trust the root cert with CT,C,C

Patch-set: 1