UserSubjectNameDefault: don't change attribute encodings
When renewing the CA cert in multiple steps, the Subject DN in the
renewed certificate uses the UTF8String attribute encoding, even if
the CSR used other attribute encodings.  This causes a variety of
problems in other programs because they may regard the new
certificate as a different CA.

The problem arises when the following procedure is used to renew a
CA certificate:

1. Generate CSR (keeping attribute encodings same as the cert)
2. Submit CSR (/profileSubmit)
3. Review request as RA agent (/profileReview)
4. Process request as RA agent (/profileProcess)

The stringified subject DN is included in the profileReview response
and re-submitted in the profileProcess request data.  Then the
string is converted back into a X500Name (this time using the
default attribute encoding i.e. UTF8String) and set into the
request, potentially changing the attribute encoding from the CSR.

To resolve the issue, when setting the Subject DN, compare the
string representations of the existing and new values.  If they are
the same, keep the old value so that the attribute encodings are
preserved.

Fixes: https://pagure.io/dogtagpki/issue/2828