Fix: #2695 Replacing Random with SecureRandom.
This is addressed by creating one centralized method to get
the SecureRandom number generator in JssSubsytem, which alread exists.

Another fix by edewata allowed us to be able to access the
JssSubsystem from most other server code. Thus only one method
required.

In a future ticket we can centralize the method calls TO the secure
random object to allows us to easily log when they are called.

Added a centralized method to CryptoUtil, which can be used by java clients.

Change-Id: I092a6c16b43ddc6bb33ff0b6084ece1f34285844