Fix issuance when CA cert lacks Subject Key ID ext
If the CA signing cert does not have the Subject Key Identifier
extension, issuance of certificates fails. Although such a CA
certificate is not compliant with RFC 5280, this does happen in the
wild, and we previously handled this case by computing the SHA-1
digest of the signing key as a last resort. This behaviour was
removed by 3c43b1119ca978c296a38a9fe404e1c0cdcdab63, breaking cert
issuance in installations with CA certs that lack the SKI extention.

Restore this behaviour.

Fixes: https://pagure.io/dogtagpki/issue/2829
Change-Id: I2f590abe258c0f9405549945b89e3c25c32c2a00