Update patch set 2
Patch Set 2:

Christina, thanks for reviewing.

 > would they care about SubjectNameDefault.java as well?  The profile
 > that uses it is caIPAserviceCert.cfg

SubjectNameDefault always populates the Subject DN from a string
(usually a pattern configured in the profile, with various attributes
from the CSR substituted into it *as strings*).  So the attribute
encoding problem doesn't arise with SubjectNameDefault (or rather,
the nature of the problem is quite different and if we ever need
to address it, the solution will look quite different).

Patch-set: 2
Label: Code-Review=0