CMSServlet.renderFinalError: log exception
renderFinalError is a "last resort" error handler that writes an
error message back to the client.  If the exception was not already
logged, the call stack will be discarded after renderFinalError is
finished with the exception.

Log the exception so that the call stack information is not lost.

Part of: https://pagure.io/dogtagpki/issue/2557

Change-Id: I2fd608adf205e3f72b67d822b1966fdb1b8bc60f
(cherry picked from commit 386357c347f8433e14ccd8637576f4c4a4e42492)