Fixed theme build script.
The theme build script has been modified to use ln command instead
of CMake create_symlink command to create links to theme files and
folders. The ln command allows the link to be created even if the
target does not exist.

This fix was based upon the previous change attempts:
* https://pagure.io/dogtagpki/issue/2582 (edewata)
* https://pagure.io/dogtagpki/issue/2687 (edewata)

https://pagure.io/dogtagpki/issue/2830

Change-Id: I4aa33f52ccec6e110b4907ac7e50f654f036a2c9