Fix #2735 Secure removal of secret data storage.
https://pagure.io/dogtagpki/issue/2735

This phase provides a method to clear memory using a configurable method, either zeroes or random bytes:

The configuration for this is:

jss.obscureMethod=random
or
jss.obscureMethod=zeroes

This fix consists of a new method in JssSubsystem and examples of the calling of this method in the kra classes. This was chosen first since it's the most crucial area where we have to obscure memory. There are cases here where secret data is being unwrapped into memory on a temporary basis. We want to clean this up as soon as possible.

The next phase will consist of the rest of the servers that do this kind of thing and some support for clients.

Also: I had to  fix  drmtest.py to complete the test run. The problem is when concatenating ints with strings in print statements: The solution is to cast the int values to a string like this str(intVar).

Fixed confusing comment in the obscureBytes method.

Change-Id: I276e3a9b678cb6b34b6c6b6a78ba2968b84f74d8