Fixed install problem in HSM case.
Previously during installation the code always tried to find all
certificates and set the trust flags. In HSM case it failed since
the temporary SSL server certificate was always stored in the
internal token and the code was looking for it in HSM.

For temporary SSL server certificate it's actually not necessary
to set the trust flags since the certificate will be replaced
with the permanent one later. So the code that sets the trust
flags has been merged into the importCert() such that it's not
executed for temporary SSL server certificate.

https://pagure.io/dogtagpki/issue/2846

Change-Id: I48b15ce9126074dc0e9fbd3dca20a615a110df0a