Fixed pki-server cert-find to work with HSM.
Previously the pki-server cert-find command would prompt for
token password if used with HSM. It has been fixed with the
following changes:

The PKISubsystem.create_subsystem_cert_object() was modified to
get the certificate info from the proper token.

The NSSDatabase.get_cert_info() was modified to specify the token
name in the certutil command if provided.

https://pagure.io/dogtagpki/issue/2901

Change-Id: If8862abe4c3057f3094c414134b9719088796963