install: support adding Subject Key ID to CSR
For externally-signed CA installation, some users want to be able to
generate a CSR with a Subject Key Identifier extension - either
user-specified or a generated default.

This commit adds support to NSSDatabase.create_request for
generating a CSR with an SKI extension.  The process to achieve this
is:

1. Generate the key.  This behaviour has been extracted to a
   separate method (NSSDatabase.generate_key).

2. If a "default" SKI is requested, generate a throw-away CSR and
   compute an SKI value from the public key contained therein.
   This is a "minimal" CSR whose only purpose is to get the public
   key in a convenient format.

3. Generate the CSR and write it to the caller-specified file.
   This CSR contains all the extensions the caller asked for.

This commit relies on an enhancement to the certutil(1) program that
allows create a CSR for a private key specified by CKA_ID.

Part-of: https://pagure.io/dogtagpki/issue/2854