install: add pkispawn option for adding SKI to CSR
For externally-signed CA installation, some users want to be able to
generate a CSR with a Subject Key Identifier extension - either
user-specified or a generated default.

This commit adds the 'pki_req_ski' pkispwan option for specifying
that the CSR should bear the SKI extension.  It can either be a
hex-encoded SKI value or the string "DEFAULT" which asks that the
value be derived from the public key.

Update the pki_default.cfg.5 man page to document the new option.

Fixes: https://pagure.io/dogtagpki/issue/2854
Change-Id: If1bf51a4935029483bba179a3f637833d0a25980